image_gen/pyproject.toml
[project]
name = "image_gen"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = ["dotenv>=0.9.9", "pillow>=11.3.0", "requests>=2.32.5"]
